refactor: tighten slack user info resolver coverage
